cdk/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.56.0
aws-cdk.aws-apigateway==1.56.0
aws-cdk.aws-certificatemanager==1.56.0
aws-cdk.aws-cloudwatch==1.56.0
aws-cdk.aws-cloudfront==1.56.0
aws-cdk.aws-codeguruprofiler==1.56.0
aws-cdk.aws-cognito==1.56.0
aws-cdk.aws-ec2==1.56.0
aws-cdk.aws-efs==1.56.0
aws-cdk.aws-events==1.56.0
aws-cdk.aws-iam==1.56.0
aws-cdk.aws-kms==1.56.0
aws-cdk.aws-lambda==1.56.0
aws-cdk.aws-lambda-event-sources==1.56.0
aws-cdk.aws-logs==1.56.0
aws-cdk.aws-s3==1.56.0
aws-cdk.aws-s3-assets==1.56.0
aws-cdk.aws-s3-deployment==1.56.0
aws-cdk.aws-sqs==1.56.0
aws-cdk.aws-ssm==1.56.0
aws-cdk.cloud-assembly-schema==1.56.0
aws-cdk.core==1.56.0
aws-cdk.cx-api==1.56.0
aws-cdk.region-info==1.56.0
cattrs==1.0.0
constructs==3.0.4
jsii==1.9.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.2
